application/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.2
Flask-RESTful==0.3.5
Flask-WTF==0.14.2
gunicorn==19.6.0
Werkzeug==0.16.1

Flask-Migrate==2.7.0

psycopg2==2.8.6
uWSGI==2.0.17.1
flask-restplus==0.13.0
requests

Flask-SQLAlchemy==2.4.1
marshmallow>=2.13.5,<3
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0

folium==0.11.0
pandas==1.1.1

dash==1.20.0
dash-bootstrap-components==0.11.1
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3